Make sure that LibreOffice is actually running before trying to stop it
This is a bit of a belt and braces commit: not only does it make the check
that LibreOffice is still running more reliable, it also adds the
--terminate_after_init option to the temporary process, making sure that it
won't be responsible for starting a new LibreOffice listener

+++ scrapy-webscanner/scanners/processors/libreoffice.py
@@ -74,12 +74,15 @@ couldn't create a LibreOffice process"""
 
     def teardown_queue_processing(self):
         if self.instance:
-            # Tell the existing instance to stop listening on the pipe (this
-            # subprocess will send that instruction and then stop immediately)
-            subprocess.run(self._make_args(accept=False))
-            # ... and *now* stop it
-            self.instance.terminate()
-            self.instance.wait()
+            if self.instance.poll() is None:
+                # Tell the existing instance to stop listening on the pipe
+                # (this subprocess will send that instruction and then stop
+                # immediately)
+                subprocess.run(
+                    self._make_args(accept=False) + ['--terminate_after_init'])
+                # ... and *now* stop it
+                self.instance.terminate()
+                self.instance.wait()
             self.instance = None
 
         # Also remove the Unix domain socket used to control access to the
